Fix issues
1. address failed when hostname is unresolved
2. no response when socks handshake failed
3. security warnings
4. unused codes

+++ urban-spork-common/src/com/urbanspork/common/protocol/vmess/header/AddressType.java
@@ -4,19 +4,19 @@ import java.util.Arrays;
 
 public enum AddressType {
 
-    IPV4(1), DOMAIN(2), IPV6(3);
+    IPV4((byte) 1), DOMAIN((byte) 2), IPV6((byte) 3);
 
-    private final int value;
+    private final byte value;
 
-    AddressType(int value) {
+    AddressType(byte value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public byte getValue() {
         return value;
     }
 
-    public static AddressType valueOf(int value) {
+    public static AddressType valueOf(byte value) {
         return Arrays.stream(values()).filter(type -> type.value == value)
             .findFirst().orElseThrow(() -> new IllegalArgumentException("Unknown address type: " + value));
     }
